--- src/App.css
@@ -1,61 +1,0 @@
-/*Page Layout*/
-.App {
-  display: flex;
-  flex-flow: row wrap;
-}
-
-.App, .App-header, .App-content, .App-sidebar, .App-footer {
-  flex: 1 100%;
-}
-
-/* Medium screens */
-@media all and (min-width: 600px) {
-  /* We tell both sidebars to share a row */
-  .App-sidebar { flex: 1 auto; }
-}
-
-/* Large screens */
-@media all and (min-width: 800px) {
-  /* We invert order of first sidebar and main
-   * And tell the main element to take twice as much width as the other two sidebars 
-   */
-  .App-content { flex: 4 0px; }
-  .App-sidebar { order: 1; }
-  .App-content { order: 2; }
-  .footer  { order: 3; }
-}
-
-/*Other Styles*/
-
-.App ul {
-  list-style: none;
-}
-.App-logo img {
-  width: 80%;
-  border-radius: 15px 50px 30px;
-  padding: 10px;
-}
-
-.App-header {
-  background-color: orange;
-  height: 100px;
-  padding: 20px;
-  color: white;
-}
-
-/*.App-container {
-  background-color: orange;
-}*/
-
-/*.App-content {
-  display: flex;
-  background-color: #fff;
-  border-radius: 50px 0 0 0;
-}
-
-.App-sidebar {
-  display: flex;
-  padding: '10px';
-  background-color: orange;
-  align-self: stretch;
-}*/
