--- src/Icon.css
@@ -1,1 +1,0 @@
-
