--- src/RR4SidebarExample.css
@@ -1,4 +1,0 @@
-/*.SidebarExample {
-float: left;
-height: 100%;
-}*/
