+++ src/Sidebar.js
@@ -7,15 +7,17 @@ class Sidebar extends Component {
   render() {
     return (
       <div>
+        <Router history={history}>
         <ul>
           <li>
-            <Icon className="App-icon" />
+              <Icon className="App-icon" />
           </li>
           <li><Link to="/">Home</Link></li>
-          <li><Link to="/hire">Hire Me</Link></li>
-          <li><Link to="/portfolio">Portfolio</Link></li>
-          <li><Link to="/about">About</Link></li>
+          <li><Link to="/#hire">Hire Me</Link></li>
+          <li><Link to="/#portfolio">Portfolio</Link></li>
+          <li><Link to="/#about">About</Link></li>
         </ul>
+        </Router>
       </div>
     );
   }
